requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
